node-data: move ledger::header in its own module
